pic2style-backend/requirements.txt
flask
flask-cors
python-dotenv
Pillow
numpy
groq